pyproject.toml
[project]
name = "legal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "google-generativeai>=0.8.6",
    "instructor>=1.14.5",
    "langgraph>=1.1.3",
    "motor>=3.7.1",
    "pdf2image>=1.17.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "uvicorn>=0.42.0",
]
